[ref #190] move remaining file store functions to extension trait
